move profile-textfield.tsx to components folder
