+++ meshapp/src/profile/profile-page.tsx
@@ -8,7 +8,7 @@ import {
   Chip,
 } from "@mui/material";
 
-import ProfileTextField from "./profile-textfield";
+import ProfileTextField from "../components/profile/profile-textfield";
 import { Profile } from "./types/profile";
 import "./styling/profile-page.css";
